+++ src/python/pants/backend/experimental/go/register.py
@@ -3,6 +3,7 @@
 
 from pants.backend.go import build, import_analysis, module, pkg, sdk, tailor, target_type_rules
 from pants.backend.go import target_types as go_target_types
+from pants.backend.go.goals import custom_goals
 from pants.backend.go.subsystems import golang
 from pants.backend.go.target_types import GoBinary, GoExternalPackageTarget, GoModule, GoPackage
 
@@ -22,4 +23,5 @@ def rules():
         *sdk.rules(),
         *tailor.rules(),
         *target_type_rules.rules(),
+        *custom_goals.rules(),
     ]
