+++ src/python/pants/backend/go/build.py
@@ -35,10 +35,9 @@ from pants.core.util_rules.source_files import SourceFiles, SourceFilesRequest
 from pants.engine.addresses import Addresses
 from pants.engine.engine_aware import EngineAwareParameter
 from pants.engine.fs import AddPrefix, CreateDigest, Digest, FileContent, MergeDigests
-from pants.engine.goal import Goal, GoalSubsystem
 from pants.engine.internals.selectors import Get, MultiGet
 from pants.engine.process import ProcessResult
-from pants.engine.rules import collect_rules, goal_rule, rule
+from pants.engine.rules import collect_rules, rule
 from pants.engine.target import (
     Dependencies,
     DependenciesRequest,
@@ -377,25 +376,6 @@ async def package_go_binary(
 
     artifact = BuiltPackageArtifact(relpath=str(output_filename))
     return BuiltPackage(digest=renamed_output_digest, artifacts=(artifact,))
-
-
-class GoBuildSubsystem(GoalSubsystem):
-    name = "go-build"
-    help = "Compile Go targets that contain source code (i.e., `go_package`)."
-
-
-class GoBuildGoal(Goal):
-    subsystem_cls = GoBuildSubsystem
-
-
-@goal_rule
-async def run_go_build(targets: UnexpandedTargets) -> GoBuildGoal:
-    await MultiGet(
-        Get(BuiltGoPackage, BuildGoPackageRequest(address=tgt.address))
-        for tgt in targets
-        if is_first_party_package_target(tgt) or is_third_party_package_target(tgt)
-    )
-    return GoBuildGoal(exit_code=0)
 
 
 def rules():
